requirements.txt
pyshark>=0.5.3
tqdm>=4.64.0
netifaces>=0.11.0
PyYAML>=6.0
